pyproject.toml
[project]
name = "blackvuesync"
version = "2.0a"
description = "Synchronize recordings from BlackVue dashcams to a local directory"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["blackvue", "dashcam", "sync", "backup"]

dependencies = []

[project.optional-dependencies]
# development dependencies
dev = [
    "behave~=1.2",
    "black~=25.11",
    "Flask~=3.1",
    "mypy~=1.14",
    "pre-commit~=4.3",
    "PyHamcrest~=2.1",
    "pytest~=8.3",
    "python-lsp-server~=1.14",
    "requests~=2.32",
    "ruff~=0.9",
]

[project.urls]
Homepage = "https://github.com/acolomba/blackvuesync"
Repository = "https://github.com/acolomba/blackvuesync"
Issues = "https://github.com/acolomba/blackvuesync/issues"

[build-system]
requires = ["setuptools>=58.0.4"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["blackvuesync"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = [
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "ARG", # flake8-unused-arguments
    "I",   # isort
    "N",   # pep8-naming
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black/ruff-format)
    "UP007", # use X | Y for type annotations (not compatible with Python 3.9)
]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["venv/", "build/", "dist/"]

[[tool.mypy.overrides]]
module = "blackvuesync_test"
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "_pytest.*"
follow_imports = "skip"

[[tool.mypy.overrides]]
module = "features.steps.*"
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "features.mock_dashcam.*"
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "requests"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "hamcrest"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "features.mock_dashcam.server"
warn_return_any = false
